+++ babel.config.js
@@ -15,4 +15,8 @@ const presets = [
     ],
 ];
 
-module.exports = { presets };
+
+module.exports = {
+    presets,
+    "plugins": ["@babel/plugin-proposal-class-properties"]
+};
