+++ package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "@babel/cli": "^7.10.1",
     "@babel/core": "^7.10.1",
+    "@babel/plugin-proposal-class-properties": "^7.10.1",
     "@babel/preset-env": "^7.10.1",
     "autoprefixer": "^9.8.0",
     "babel-loader": "^8.1.0",
