+++ src/index.html
@@ -6,12 +6,11 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
   <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.css %>">
-  <link rel="stylesheet" href="./pages/index.css">
 </head>
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg').default%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -31,20 +30,12 @@
 
     <div class="popup">
       <div class="field-not-clickable" id="popup-content">
-        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close new-card__close">
+        <img src="<%=require('./images/close.svg').default%>" alt="" class="popup__close new-card__close">
       </div>
     </div>
     
   </div>
 
-  <script src="./script/api.js"></script>
-  <script src="./script/initialCards.js"></script>
-  <script src="./script/card.js"></script>
-  <script src="./script/cardList.js"></script>
-  <script src="./script/popup.js"></script>
-  <script src="./script/userInfo.js"></script>
-  <script src="./script/formValidator.js"></script>
-  <script src="./script/script.js"></script>
   <script src="<%= htmlWebpackPlugin.files.js %>"></script>
 
 </body>
