+++ src/script/api.js
@@ -1,4 +1,4 @@
-class Api {
+export default class Api {
     constructor(options) {
       this.options = options;
       this.url = this.options.baseUrl;
