+++ src/style.css
@@ -1,3 +1,1 @@
-h1 {
-    color: red;
-}
+@import "./pages/index.css";
