+++ webpack.config.js
@@ -39,7 +39,7 @@ module.exports = {
             {
                 test: /\.(png|jpg|gif|ico|svg)$/,
                 use: [
-                    'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                    'file-loader?name=./images/[name].[ext]', // указали папку, куда складывать изображения
                     {
                         loader: 'image-webpack-loader',
                         options: {}
